requirements.txt
exceptiongroup==1.2.2
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
pytest==8.3.5
python-dotenv==1.1.0
tomli==2.2.1
flake8
pytest
allure-pytest
